requirement.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pdfkit==0.6.1
Pillow==2.2.1
pkg-resources==0.0.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
